requirements.txt
pydoll-python
streamlit